backend/requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.1
dill==0.3.7
Django==3.2
django-enumfield==3.1
django-filter==23.2
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
django-types==0.17.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-writable-nested==0.7.0
envkey==2.4.2
gunicorn==21.2.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
nose==1.3.7
olefile==0.46
packaging==23.1
phonenumbers==8.13.19
Pillow==10.0.0
platformdirs==3.10.0
psycopg==3.1.10
psycopg2-binary==2.9.7
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
